requirements.txt
Flask==1.1.4
Flask-Cors
Flask-Migrate==2.7.0
Flask-SocketIO==4.3.2
Flask-Sockets
Flask-SQLAlchemy
Flask-WTF
wtforms_json
itsdangerous==1.1.0
termcolor
gunicorn
